LayoutTests/fast/mediastream/RTCPeerConnection-state.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
    </head>
    <body>
        <script>
            description("Tests the RTCPeerConnection states.");

            var pc = null;

            function stateChanged()
            {
                testPassed("stateChanged was called");
                shouldBeEqualToString('pc.signalingState', 'closed');
                finishJSTest();
            }

            shouldNotThrow("pc = new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}]});");
            shouldBeEqualToString('pc.signalingState', 'stable');
            pc.onsignalingstatechange = stateChanged;
            pc.close();

            window.jsTestIsAsync = true;
            window.successfullyParsed = true;
        </script>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
